refactor: Grouped some routes so that their layout doesn't affect the chats route
